v10.18: Precision hardening — double precision in 4 more locations
Fixed precision in:
1. dn_l2norm: float32 → double for sum-of-squares
2. gqa_attention Q@K dot: float32 → double accumulation
3. Shared expert gate dot: AVX float32 → double scalar
4. Quantum physics prompt for harder testing

Output for "Hello" prompt produces English words (Cancel, dismiss).
Output for "Explain quantum physics" still produces single chars (V, A, .).
This suggests the DeltaNet state doesn't properly accumulate information
across longer prompts — the context understanding is weak.

The remaining gap is likely in how the DeltaNet recurrence preserves
information across tokens, not in per-token precision.

+++ engine/runtime/attention.h
@@ -150,10 +150,10 @@ static void gqa_attention(
         int t;
         for (t = 0; t < seq_len; t++) {
             const float* kt = kv->keys + t * num_kv_heads * head_dim + kv_head * head_dim;
-            float dot = 0.0f;
+            double dot = 0.0;
             int d;
-            for (d = 0; d < head_dim; d++) dot += qh[d] * kt[d];
-            scores[t] = dot * scale;
+            for (d = 0; d < head_dim; d++) dot += (double)qh[d] * (double)kt[d];
+            scores[t] = (float)(dot * scale);
         }
 
         /* Softmax */
